requirements.txt
aiogram
python-dotenv
aioschedule